+++ include/igraph_constants.h
@@ -206,6 +206,11 @@ typedef enum { IGRAPH_VORONOI_FIRST = 0,
                IGRAPH_VORONOI_RANDOM
              } igraph_voronoi_tiebreaker_t;
 
+typedef enum { IGRAPH_CHUNG_LU_ORIGINAL = 0,
+               IGRAPH_CHUNG_LU_GRG,
+               IGRAPH_CHUNG_LU_NR
+             } igraph_chung_lu_t;
+
 typedef enum { IGRAPH_ROW_MAJOR = 0,
                IGRAPH_COLUMN_MAJOR = 1
              } igraph_matrix_storage_t;